backend/requirements.txt
# Web Framework
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.21
watchfiles==1.1.1

# AI / ML
google-genai==1.59.0
opencv-python-headless==4.13.0.90
onnxruntime==1.23.2

# Database / Storage
supabase==2.27.2

# Data / Schema Validation
pydantic==2.12.5
pydantic-settings==2.12.0

# Utils
python-dotenv==1.2.1
